Added in functionality for opportunity components.

+++ src/java/org/aspcfs/modules/accounts/actions/Accounts.java
@@ -1060,14 +1060,18 @@ public final class Accounts extends CFSModule {
     try {
       db = this.getConnection(context);
       thisOrg = new Organization(db, Integer.parseInt(id));
+      htmlDialog.setTitle("CFS: Account Management");
       htmlDialog.setRelationships(thisOrg.processDependencies(db));
-
-      htmlDialog.setTitle("CFS: Confirm Delete");
-      htmlDialog.setHeader("The account you are requesting to delete has the following dependencies within CFS:");
-      htmlDialog.addButton("Delete All", "javascript:window.location.href='/Accounts.do?command=Delete&action=delete&orgId=" + thisOrg.getOrgId() + "'");
-      htmlDialog.addButton("Disable Only", "javascript:window.location.href='/Accounts.do?command=Delete&orgId=" + thisOrg.getOrgId() + "&action=disable'");
-      htmlDialog.addButton("Cancel", "javascript:parent.window.close()");
-
+      
+      if (thisOrg.getHasOpportunities()) {
+          htmlDialog.setHeader("Please re-assign or delete any opportunities associated with this account first.");
+          htmlDialog.addButton("OK", "javascript:parent.window.close()");
+      } else {
+        htmlDialog.setHeader("The account you are requesting to delete has the following dependencies within CFS:");
+        htmlDialog.addButton("Delete All", "javascript:window.location.href='/Accounts.do?command=Delete&action=delete&orgId=" + thisOrg.getOrgId() + "'");
+        htmlDialog.addButton("Disable Only", "javascript:window.location.href='/Accounts.do?command=Delete&orgId=" + thisOrg.getOrgId() + "&action=disable'");
+        htmlDialog.addButton("Cancel", "javascript:parent.window.close()");
+      }
     } catch (Exception e) {
       errorMessage = e;
     } finally {
